rev scripts

+++ adapt_experiments/background_adapt.sh
@@ -2,7 +2,11 @@
 
 PID_DIR=/jetstream
 
+
+if [ -f ${PID_DIR}/shaping.pid ]; then
+echo "stopping previous run"
 kill `cat ${PID_DIR}/shaping.pid`
+fi
 
 nohup python /jetstream/js/adapt_experiments/wifi_tc_simulate.py > /jetstream/shape.out  2>&1 $@ &
 PID=$!
